Move geFormElementHelper to own file, revert uneeded changes / cleanup
