+++ src/lwc/geTemplateBuilderFormFields/geTemplateBuilderFormFields.html
@@ -259,6 +259,7 @@
                                                             active-form-section-id={activeFormSectionId}
                                                             is-first={it.first}
                                                             is-last={it.last}
+                                                            source-object-fields-describe={sourceObjectFieldsDescribe}
                                                             onchangeactivesection={handleChangeActiveSection}
                                                             onformsectionup={handleFormSectionUp}
                                                             onformsectiondown={handleFormSectionDown}
